requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
importlib-resources==5.10.2
jsonschema==4.17.3
pkgutil-resolve-name==1.3.10
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.19.3
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
zipp==3.13.0
